Update folders and csproj to match other building blocks
