--- actors/csharp/sdk/smartdevice.Service/smartdevice.Service.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Dapr.Actors" Version="1.9.0" />
-    <PackageReference Include="Dapr.Actors.AspNetCore" Version="1.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\smartdevice.Interfaces\smartdevice.Interfaces.csproj" />
-  </ItemGroup>
-
-</Project>
